Fix the type of the first argument of the utils::inb function

+++ content/src/utils.rs
@@ -6,7 +6,7 @@ pub fn outb(port: u16, val: u8) {
     }
 }
 
-pub fn inb(port: u8) -> u8 {
+pub fn inb(port: u16) -> u8 {
     let result: u8;
 
     unsafe {
